requirements.txt
black=21.11b1
Jinja2==3.0.1
loguru==0.5.3
pandas==1.3.4
pycodestyle=2.8.0
python-dotenv==0.19.2
sendgrid==6.7.1
tweepy==3.10.0